receivers: pass emails through net/mail.ParseAddress
This performs a basic format check on the provided email address and
strips out extra junk that RFC 5322 allows. (i.e. 'Bob <bob@example.com')

We don't validate the hostname or mailbox (by sending an activation
email) and I'm unsure if this should be done in a new service or
not. User signup needs to validate emails also, so it seems either
Auth should perform that task or it should be left up to another service.