files: drop cpu usage (infinite loop) in transfer merging for upload
This was reported by Ray as he noticed high CPU usage when paygate was
idle shortly after startup. A quick profile shows this looping was the
main cause as it didn't properly exit out in error conditions.

For now we're going to remove the errCount condition and just error
whenever a problem is encountered. If lower levels need retries then
we can add those on the lower level.