transfers: expand window for EffectiveEntryDate comparison against created_at
This issue was reported by Ray regarding same day transfers, but the
logic applies as Transfers could be created on a weekend and not
posted for a day or two (which could push outside of the window).

Also, if a transfer is created today paygate sets the
EffectiveEntryDate to tomorrow in some cases and thus if the file is
immediately returned would fall outside of today's window.