internal/filetransfer: initial setup for handling incoming NOC/COR files
These files can give corrected data for an EntryDetail (which might
allow us to fix a Depository's data) that was sent back after we
uploaded a file. I'm not sure if we'll be able to automatically fixup
any data, but these files do we need to be handled by an ACH originator.