all: split 'err and object nil' checks into two blocks
This error shows itself as returning "200 OK" in HTTP routes when
there's an underlying error. The error is logged, but the route does
not return an error.

The way this works is we call moovhttp.Prolbem(nil) which skips over
setting the status code to http.StatusBadRequest and any error
marshaling. Then the http.ResponseWriter will assume "200 OK" when the
response finishes.