transfers/pipeline: write merged files to disk
This allows us to save them for a period of time and audit what was
sent off. Ideally we need to name them the same as what they're
uploaded as, but that's not quite easy with how the code is setup now.

Also, these need to be encrypted and stored somewhere (cloud storage).